requirements/runtime.txt
# Add requirements here, use the script for help (erotemic needs to publish this script)
# python ~/local/tools/supported_python_versions_pip.py pygments
# python ~/local/tools/supported_python_versions_pip.py ubelt
# python ~/local/tools/supported_python_versions_pip.py scriptconfig
# python ~/local/tools/supported_python_versions_pip.py toml
# python ~/local/tools/supported_python_versions_pip.py rich

xdev>=0.3.1     ;                            python_version >= '3.6'    # Python 3.6+

ubelt>=1.2.1     ;                            python_version >= '3.6'    # Python 3.6+
ubelt>=0.8.1     ; python_version < '3.6' and python_version >= '2.7'    # Python 2.7

toml>=0.9.6     ;                            python_version >= '2.7'    # Python 2.7+

scriptconfig>=0.6.0     ;                            python_version >= '3.6'    # Python 3.6+
scriptconfig>=0.5.0     ; python_version < '3.6' and python_version >= '2.7'    # Python 2.7

rich>=12.4.4    ;                            python_version >= '3.8'    # Python 3.8+
rich>=11.2.0    ; python_version < '3.8' and python_version >= '3.6'    # Python 3.6 

git-python>=1.0.0     ;                            python_version >= '2.7'    # Python 2.7+

black>=22.1.0    ;                            python_version >= '3.6.2'    # Python 3.6.2+

packaging>=21.3

PyYAML >= 0.6.0
ruamel.yaml >= 0.17.21
